Fix log format strings

+++ internal/impl/opensearch/output.go
@@ -347,10 +347,10 @@ func (e *Output) WriteBatch(ctx context.Context, msg service.MessageBatch) error
 	dur := time.Since(start)
 
 	e.log.Debugf(
-		"Successfully dispatched [%d] documents in %s (%f docs/sec)",
+		"Successfully dispatched [%d] documents in %s (%.2f docs/sec)",
 		biStats.NumFlushed,
 		dur.Truncate(time.Millisecond),
-		int64(1000.0/float64(dur/time.Millisecond)*float64(biStats.NumFlushed)),
+		1000.0/float64(dur/time.Millisecond)*float64(biStats.NumFlushed),
 	)
 	return nil
 }
